Fix a display bugs on wti rm command

+++ lib/web_translate_it/command_line.rb
@@ -135,20 +135,33 @@ module WebTranslateIt
       WebTranslateIt::Connection.new(configuration.api_key) do |http|
         parameters.each do |param|
           if Util.ask_yes_no("Are you sure you want to delete the master file #{param}?\nThis will also delete its target files and translations.", false)
-            configuration.files.find_all{ |file| file.file_path == param }.each do |master_file|
-              master_file.delete(http)
-              # delete files
-              success = File.delete(master_file.file_path) if File.exists?(master_file.file_path)
-              complete_success = false if !success
-              configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
-                success = File.delete(target_file.file_path) if File.exists?(target_file.file_path)
+            files = configuration.files.find_all{ |file| file.file_path == param }
+            if files.any?
+              files.each do |master_file|
+                master_file.delete(http)
+                # delete files
+                if File.exists?(master_file.file_path)
+                  success = File.delete(master_file.file_path)
+                  puts StringUtil.success("Deleted master file #{master_file.file_path}.") if success
+                end
                 complete_success = false if !success
+                configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
+                  if File.exists?(target_file.file_path)
+                    success = File.delete(target_file.file_path)
+                    puts StringUtil.success("Deleted target file #{target_file.file_path}.") if success
+                  else
+                    puts StringUtil.failure("Target file #{target_file.file_path} doesn’t exist locally")
+                  end
+                  complete_success = false if !success
+                end
               end
+              puts StringUtil.success("All done.") if complete_success
+            else
+              puts StringUtil.failure("#{param}: File doesn’t exist on project.")
             end
           end
         end
       end
-      puts StringUtil.success("Master file deleted.")
       complete_success
     end
